comm

+++ client/src/app/Notification/notifications/notifications.component.ts
@@ -1,23 +1,23 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { NotificationsService } from '../../services/notifications.service';
-
 import { FormsModule } from '@angular/forms';
 import { Notification } from '../../models/notification';
 
 @Component({
   selector: 'app-notifications',
+  standalone: true,
   imports: [
     RouterModule,
     FormsModule,
     CommonModule
   ],
   templateUrl: './notifications.component.html',
-  styleUrl: './notifications.component.css'
+  styleUrls: ['./notifications.component.css']
 })
-export class NotificationsComponent {
-  notifications: Notification []=[];
+export class NotificationsComponent implements OnInit {
+  notifications: Notification[] = [];
   searchText: string = '';
 
   constructor(private notificationService: NotificationsService) {}
@@ -27,9 +27,12 @@ export class NotificationsComponent {
       this.notifications = data;
     });
   }
-  delete(maThongBao: string){
-    this.notificationService.deleteNotification(maThongBao).subscribe(() => {
-      this.ngOnInit();
-    });
+
+  delete(maThongBao: string): void {
+    if (confirm('Bạn có chắc muốn xoá thông báo này không?')) {
+      this.notificationService.deleteNotification(maThongBao).subscribe(() => {
+        this.ngOnInit();
+      });
+    }
   }
 }
